tidy up lightbox module

+++ modules_v2/lightbox/functions/lightbox_print_media_row.php
@@ -2,10 +2,10 @@
 /**
  * Lightbox Album module for webtrees
  *
- * Display media Items using Lightbox 4.1
+ * Display media Items using Lightbox
  *
  * webtrees: Web based Family History software
- * Copyright (C) 2010 webtrees development team.
+ * Copyright (C) 2011 webtrees development team.
  *
  * Derived from PhpGedView
  * Copyright (C) 2007 to 2009  PGV Development Team.  All rights reserved.
@@ -24,8 +24,6 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * @package webtrees
- * @subpackage Module
  * @version $Id$
  * @author Brian Holland
  */
@@ -43,43 +41,39 @@ if (!defined('WT_WEBTREES')) {
  */
 function lightbox_print_media_row($rtype, $rowm, $pid) {
 
-	global $WT_IMAGES, $MEDIA_DIRECTORY, $TEXT_DIRECTION;
-	global $GEDCOM, $THUMBNAIL_WIDTH, $USE_MEDIA_VIEWER;
-	global $SEARCH_SPIDER;
-	global $t, $n, $item, $items, $p, $edit, $reorder, $LB_AL_THUMB_LINKS, $note;
-	global $LB_URL_WIDTH, $LB_URL_HEIGHT, $order1, $sort_i, $notes, $q, $LB_TT_BALLOON;
+	global $TEXT_DIRECTION, $LB_TT_BALLOON;
+	global $item, $sort_i, $notes;
 
 	$reorder=safe_get('reorder', '1', '0');
 
-	$mainMedia = check_media_depth($rowm["m_file"], "NOTRUNC");
+	$mainMedia = check_media_depth($rowm['m_file'], 'NOTRUNC');
 	// If media file is missing from "media" directory, but is referenced in Gedcom
 	if (!media_exists($mainMedia)) {
 		if (!file_exists($rowm['m_file']) && !isset($rowm['m_file'])) {
-			echo "<tr>";
-				echo "<td valign=\"top\" rowspan=\"2\" >";
-					echo '<img src="', WT_MODULES_DIR, 'lightbox/images/transp80px.gif" height="82px" alt=""></img>';
-				echo "</td>";
-				echo "<td class=\"description_box\" valign=\"top\" colspan=\"3\" nowrap=\"nowrap\" >";
-					echo "<center><br /><img src=\"" . WT_THEME_DIR . "images/media.gif\" height=\"30\" border=\"0\" />";
-					echo "<font size=\"1\"><br />" . WT_I18N::translate('File not found.') . "</font></center>";
-				echo "</td>";
-			echo "</tr>";
+			echo '<tr>';
+			echo '<td valign="top" rowspan="2" >';
+			echo '<img src="', WT_MODULES_DIR, 'lightbox/images/transp80px.gif" height="82px" alt=""></img>';
+			echo '</td>';
+			echo '<td class="description_box" valign="top" colspan="3" nowrap="nowrap">';
+			echo '<center><br /><img src="', WT_THEME_DIR, 'images/media.gif" height="30" border="0" />';
+			echo '<font size="1"><br />', WT_I18N::translate('File not found.'), '</font></center>';
+			echo '</td>';
+			echo '</tr>';
 		} else if (!file_exists($rowm['m_file'])) {
-			echo "<li class=\"li_norm\" >";
-			echo "<table class=\"pic\" width=\"50px\" border=\"0\" >";
-			echo "<tr>";
-				echo "<td valign=\"top\" rowspan=\"2\" >";
-					echo '<img src="', WT_MODULES_DIR, 'lightbox/images/transp80px.gif" height="100px" alt=""></img>';
-				echo "</td>";
-				echo "<td class=\"description_box\" valign=\"top\" colspan=\"3\" nowrap=\"nowrap\" >";
-					echo "<center><br /><img src=\"" . WT_THEME_DIR . "images/media.gif\" height=\"30\" border=\"0\" />";
-					echo "<font size=\"1\"><br />" . WT_I18N::translate('File not found.') . "</font></center>";
-				echo "</td>";
-				echo "</tr>";
-
+			echo '<li class="li_norm" >';
+			echo '<table class="pic" width="50px" border="0" >';
+			echo '<tr>';
+			echo '<td valign="top" rowspan="2" >';
+			echo '<img src="', WT_MODULES_DIR, 'lightbox/images/transp80px.gif" height="100px" alt=""></img>';
+			echo '</td>';
+			echo '<td class="description_box" valign="top" colspan="3" nowrap="nowrap" >';
+			echo '<center><br /><img src="', WT_THEME_DIR, 'images/media.gif" height="30" border="0" />';
+			echo '<font size="1"><br />', WT_I18N::translate('File not found.'), '</font></center>';
+			echo '</td>';
+			echo '</tr>';
 		} else {
-			echo "<li class=\"li_norm\" >";
-			echo "<table class=\"pic\" width=\"50px\" border=\"0\" >";
+			echo '<li class="li_norm" >';
+			echo '<table class="pic" width="50px" border="0" >';
 		}
 	// Else Media files are present in "media" directory
 	} else {
@@ -90,32 +84,32 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 			// Media is NOT linked to private person
 			// If reorder media has been clicked
 			if (isset($reorder) && $reorder==1) {
-				echo "<li class=\"facts_value\" style=\"border:0px;\" id=\"li_" . $rowm['m_media'] . "\" >";
+				echo '<li class="facts_value" style="border:0px;" id="li_', $rowm['m_media'], '" >';
 
 			// Else If reorder media has NOT been clicked
 			// Highlight Album Thumbnails - Changed=new (blue), Changed=old (red), Changed=no (none)
 			} else if ($rtype=='new') {
-				echo "<li class=\"li_new\">";
+				echo '<li class="li_new">';
 			} else if ($rtype=='old') {
-				echo "<li class=\"li_old\">";
+				echo '<li class="li_old">';
 			} else {
-				echo "<li class=\"li_norm\">";
+				echo '<li class="li_norm">';
 			}
 		}
 	}
 
 	// Add blue or red borders
-	$styleadd="";
-	if ($rtype=='new') $styleadd = "change_new";
-	if ($rtype=='old') $styleadd = "change_old";
+	$styleadd='';
+	if ($rtype=='new') $styleadd = 'change_new';
+	if ($rtype=='old') $styleadd = 'change_old';
 
 	// NOTE Start printing the media details
 	if (!media_exists($mainMedia)) {
 		if (!media_exists($rowm['m_file'])) {
-			$thumbnail = "";
-			$isExternal = ""; // isFileExternal($thumbnail);
+			$thumbnail = '';
+			$isExternal = ''; // isFileExternal($thumbnail);
 		} else {
-			$thumbnail = thumbnail_file($rowm["m_file"], true, false, $pid);
+			$thumbnail = thumbnail_file($rowm['m_file'], true, false, $pid);
 			$isExternal = isFileExternal($thumbnail);
 		}
 	} else {
@@ -129,44 +123,44 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 	if (canDisplayFact($pid, $rowm['m_file'], $rowm['mm_gedrec'])) {
 
 		//  Get the title of the media
-		$media=WT_Media::getInstance($rowm["m_media"]);
-		$rawTitle = $rowm["m_titl"];
-		if (empty($rawTitle)) $rawTitle = get_gedcom_value("TITL", 2, $rowm["mm_gedrec"]);
-		if (empty($rawTitle)) $rawTitle = basename($rowm["m_file"]);
+		$media=WT_Media::getInstance($rowm['m_media']);
+		$rawTitle = $rowm['m_titl'];
+		if (empty($rawTitle)) $rawTitle = get_gedcom_value('TITL', 2, $rowm['mm_gedrec']);
+		if (empty($rawTitle)) $rawTitle = basename($rowm['m_file']);
 		$mediaTitle = PrintReady(htmlspecialchars($rawTitle));
 
-		$mainMedia = check_media_depth($rowm["m_file"], "NOTRUNC");
+		$mainMedia = check_media_depth($rowm['m_file'], 'NOTRUNC');
 		$mainFileExists = true;
 		$imgsize = findImageSize($mainMedia);
 		$imgwidth = $imgsize[0]+40;
 		$imgheight = $imgsize[1]+150;
 
 		// Get the tooltip link for source
-		$sour = get_gedcom_value("SOUR", 1, $rowm["m_gedrec"]);
+		$sour = get_gedcom_value('SOUR', 1, $rowm['m_gedrec']);
 
 		//Get media item Notes
-		$haystack = $rowm["m_gedrec"];
-		$needle   = "1 NOTE";
+		$haystack = $rowm['m_gedrec'];
+		$needle   = '1 NOTE';
 		$before   = substr($haystack, 0, strpos($haystack, $needle));
 		$after    = substr(strstr($haystack, $needle), strlen($needle));
 		$final    = $before.$needle.$after;
 		$notes    = PrintReady(htmlspecialchars(print_fact_notes($final, 1, true, true)));
 
 		// Get info on how to handle this media file
-		$mediaInfo = mediaFileInfo($mainMedia, $thumbnail, $rowm["m_media"], $mediaTitle, $notes);
+		$mediaInfo = mediaFileInfo($mainMedia, $thumbnail, $rowm['m_media'], $mediaTitle, $notes);
 
 		//text alignment for Tooltips
-		if ($TEXT_DIRECTION=="rtl") {
-			$alignm = "right";
-			$left = "true";
+		if ($TEXT_DIRECTION=='rtl') {
+			$alignm = 'right';
+			$left = 'true';
 		} else {
-			$alignm = "left";
-			$left = "false";
+			$alignm = 'left';
+			$left = 'false';
 		}
 
 		// Tooltip Options
 		$tt_opts = ", BALLOON," . $LB_TT_BALLOON ;
-		$tt_opts .= ", LEFT," . $left . "";
+		$tt_opts .= ", LEFT," . $left;
 		$tt_opts .= ", ABOVE, true";
 		$tt_opts .= ", TEXTALIGN, '" . $alignm . "'";
 		$tt_opts .= ", WIDTH, -480 ";
@@ -187,12 +181,12 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 		$tt_opts .= ", DELAY, 0";
 
 		// Prepare Below Thumbnail  menu ----------------------------------------------------
-		if ($TEXT_DIRECTION== "rtl") {
-			$submenu_class = "submenuitem_rtl";
-			$submenu_hoverclass = "submenuitem_hover_rtl";
+		if ($TEXT_DIRECTION== 'rtl') {
+			$submenu_class = 'submenuitem_rtl';
+			$submenu_hoverclass = 'submenuitem_hover_rtl';
 		} else {
-			$submenu_class = "submenuitem";
-			$submenu_hoverclass = "submenuitem_hover";
+			$submenu_class = 'submenuitem';
+			$submenu_hoverclass = 'submenuitem_hover';
 		}
 		$menu = new WT_Menu();
 		// Truncate media title to 13 chars (45 chars if Streetview) and add ellipsis
@@ -205,26 +199,26 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 		$mtitle = PrintReady(htmlspecialchars($mtitle));
 
 		// Continue menu construction
-		// If media file is missing from "media" directory, but is referenced in Gedcom
+		// If media file is missing from 'media' directory, but is referenced in Gedcom
 		if (!media_exists($rowm['m_file']) && !media_exists($mainMedia)) {
-			$menu->addLabel("<img src=\"{$thumbnail}\" style=\"display:none;\" alt=\"\" title=\"\" />" . WT_I18N::translate('Edit')." (". $rowm["m_media"].")", "right");
+			$menu->addLabel("<img src=\"{$thumbnail}\" style=\"display:none;\" alt=\"\" title=\"\" />" . WT_I18N::translate('Edit')." (". $rowm['m_media'].")", 'right');
 		} else {
-			$menu->addLabel("<img src=\"{$thumbnail}\" style=\"display:none;\" alt=\"\" title=\"\" />" . PrintReady($mtitle), "right");
+			$menu->addLabel("<img src=\"{$thumbnail}\" style=\"display:none;\" alt=\"\" title=\"\" />" . PrintReady($mtitle), 'right');
 		}
 		// Next line removed to avoid gallery thumbnail duplication
-		// $menu["link"] = mediaInfo['url'];
+		// $menu['link'] = mediaInfo['url'];
 
 		if ($rtype=='old') {
 			// Do not print menu if item has changed and this is the old item
 		} else {
 			// Continue printing menu
-			$menu->addClass("", "", "submenu");
+			$menu->addClass('', '', 'submenu');
 
 			// View Notes
 			if (strpos($rowm['m_gedrec'], "\n1 NOTE")) {
-				$submenu = new WT_Menu("&nbsp;&nbsp;" . WT_I18N::translate('View Notes') . "&nbsp;&nbsp;", "#", "right");
+				$submenu = new WT_Menu('&nbsp;&nbsp;' . WT_I18N::translate('View Notes') . '&nbsp;&nbsp;', '#', 'right');
 				// Notes Tooltip ----------------------------------------------------
-				$sonclick  = "TipTog(";
+				$sonclick  = 'TipTog(';
 				// Contents of Notes
 				$sonclick .= "'";
 				$sonclick .= "&lt;font color=#008800>&lt;b>" . WT_I18N::translate('Notes') . ":&lt;/b>&lt;/font>&lt;br />";
@@ -239,7 +233,7 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 				$menu->addSubMenu($submenu);
 			}
 			//View Details
-			$submenu = new WT_Menu("&nbsp;&nbsp;" . WT_I18N::translate('View Details') . "&nbsp;&nbsp;", WT_SERVER_NAME.WT_SCRIPT_PATH . "mediaviewer.php?mid=" . $rowm["m_media"], "right");
+			$submenu = new WT_Menu("&nbsp;&nbsp;" . WT_I18N::translate('View Details') . "&nbsp;&nbsp;", WT_SERVER_NAME.WT_SCRIPT_PATH . "mediaviewer.php?mid=" . $rowm['m_media'], "right");
 			$submenu->addClass($submenu_class, $submenu_hoverclass);
 			$menu->addSubMenu($submenu);
 			//View Source
@@ -353,34 +347,26 @@ function lightbox_print_media_row($rtype, $rowm, $pid) {
 				echo "<td width=\"5px\"></td>";
 				echo "</tr>";
 			}
-
 			// echo "</table>";
 		}
-
 	} // NOTE End If Show fact details
-
 
 	// If media file is missing but details are in Gedcom then add the menu as well
 	//if (!media_exists($rowm['m_file'])) {
-	if (!media_exists($mainMedia)) {
-		if (!media_exists($rowm['m_file'])) {
-			echo "<tr>";
-			echo "<td ></td>";
-			echo "<td valign=\"bottom\" align=\"center\" nowrap=\"nowrap\">";
-			echo $menu->getMenu();
-			echo "</td>";
-			echo "<td ></td>";
-			echo "</tr>";
-		}
+	if (!media_exists($mainMedia) && !media_exists($rowm['m_file'])) {
+		echo '<tr>';
+		echo '<td ></td>';
+		echo '<td valign="bottom" align="center" nowrap="nowrap">';
+		echo $menu->getMenu();
+		echo '</td>';
+		echo '<td ></td>';
+		echo '</tr>';
 	}
-
 	//close off the table
-	echo "</table>";
-
+	echo '</table>';
 	$media_data = $rowm['m_media'];
-	echo "<input type=\"hidden\" name=\"order1[$media_data]\" value=\"$sort_i\" />";
+	echo '<input type="hidden" name="order1[', $media_data, ']" value="', $sort_i, '" />';
 	$sort_i++;
-
-	echo "</li>";
+	echo '</li>';
 	return true;
 }
